Refactor scheduling utilities into separate modules
