+++ app/__init__.py
@@ -9,7 +9,8 @@ import psutil
 from flask import Flask
 
 from app.utils.retention_policy import retention_cleanup
-from app.utils.scheduling import schedule_crawlers, schedule_summarization, scheduler, start_log_caching
+from app.utils.scheduler import schedule_crawlers, schedule_summarization, scheduler
+from app.utils.scheduling import start_log_caching
 from app.utils.video_archiver import archive_screenshots, compile_to_teaser
 from app.config import backup_config, restore_config
 from app.utils.email_alerts import email_alert
@@ -169,7 +170,7 @@ def create_app(watchdog=True, schedule=True):
         app.watchdog_thread = watchdog_thread
 
     # Start collecting metrics
-    from .utils.scheduling import start_metrics_collection
+    from .utils.metrics import start_metrics_collection
     start_metrics_collection()
 
     start_log_caching()
